requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
pytest==7.4.3
black==23.12.1
mypy==1.8.0
pylint==3.0.3 